backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["anderskloborg <anderskloborg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
openai = "^0.28.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.209.5"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
asyncpg = "^0.28.0"
gunicorn = "^21.2.0"
boto3 = "^1.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
